% Generated by roxygen2: do not edit by hand
\name{as.big.data.table}
\alias{as.big.data.table}
\alias{as.big.data.table.call}
\alias{as.big.data.table.data.table}
\alias{as.big.data.table.function}
\alias{as.big.data.table.list}
\title{Convert to big.data.table}
\usage{
as.big.data.table(x, ...)

\method{as.big.data.table}{function}(x, rscl, partition.by, partitions,
  parallel = TRUE, ...)

\method{as.big.data.table}{call}(x, rscl, partition.by, partitions,
  parallel = TRUE, ...)

\method{as.big.data.table}{list}(x, partition.by, partitions, parallel = TRUE,
  ...)

\method{as.big.data.table}{data.table}(x, rscl, partition.by, partitions,
  parallel = TRUE, ...)
}
\arguments{
\item{x}{object to cast into big.data.table, can be data.table, function, quoted call or a list.}

\item{rscl}{list of Rserve connections, can be validated by `is.rscl(rscl, silent=FALSE)`.}

\item{partition.by}{character vector of column names to be used for partitioning, `uniqueN` by those columns should be lower than number of nodes.}

\item{partitions}{data.table of unique combinations of values in *partition.by* columns.}

\item{parallel}{logical, see `?bdt.eval`.}

\item{\dots}{arguments passed to methods.}
}
\value{
big.data.table object.
}
\description{
Convert to big.data.table
}
\note{
Supported `x` data types are *data.table* (will be automatically spread across the nodes), *function* or quoted *call* will be evaluated on each node and assigned to `x` variable, a *list* will struct big.data.table on already working set of nodes, all having `x` data.tables.
}

