% Generated by roxygen2: do not edit by hand
\name{as.data.table.big.data.table}
\alias{as.data.table.big.data.table}
\title{Extracting data from nodes to local R session}
\usage{
\method{as.data.table}{big.data.table}(x, ...,
  .log = getOption("bigdatatable.log", FALSE))
}
\arguments{
\item{x}{big.data.table}

\item{.log}{logical if *TRUE* then logging will be done using logR to postgres db.}

\item{\dots}{arguments passed to `bdt.eval`.}
}
\value{
data.table
}
\description{
Extracting data from nodes to local R session
}

