% Generated by roxygen2: do not edit by hand
\name{bdt.eval}
\alias{bdt.eval}
\title{big.data.table evaluate on nodes}
\usage{
bdt.eval(x, expr, qexpr, send = FALSE, simplify = TRUE, rbind = TRUE,
  try = FALSE, parallel = TRUE)
}
\arguments{
\item{x}{big.data.table.}

\item{expr}{expression.}

\item{qexpr}{quoted expression.}

\item{send}{logical, if TRUE submit expression appended with `TRUE` to not fetch potentially big results from provided *expr*, useful for data.table *set** or `:=` functions.}

\item{simplify}{logical if *TRUE* (default) it will simplify list of 1 length same type objects to vector.}

\item{rbind}{logical if *TRUE* (default) results are data.table they will be rbinded.}

\item{try}{should be wrapped into tryCatch for errors and warnings? default (TRUE), means that warnings will not return the actual value.}

\item{parallel}{logical if parallel *TRUE* (default) it will send expression to nodes using `wait=FALSE` and collect results afterward executing each node in parallel.}
}
\value{
Depending on *simplify, rbind* the results of evaluated expression.
}
\description{
Main engine for passing queries to nodes, control parallelization, rbinding or simplifing returned object. Allows to measure timing and verbose messages.
}

