% Generated by roxygen2: do not edit by hand
\name{bdt.partition}
\alias{bdt.partition}
\title{Controls partitioning}
\usage{
bdt.partition(x, partition.by, parallel = TRUE,
  .log = getOption("bigdatatable.log", FALSE))
}
\arguments{
\item{x}{big.data.table.}

\item{partition.by}{character.}

\item{parallel,}{see `bdt.eval`.}

\item{.log}{passed to `bdt.eval`.}
}
\value{
big.data.table object.
}
\description{
By default function will validate if data partitioning is enforced on the nodes.
}

