% Generated by roxygen2: do not edit by hand
\name{clean}
\alias{clean}
\title{Force collect results to unlock error}
\usage{
clean(rscl)
}
\arguments{
\item{rscl}{list of connections to R nodes}
}
\value{
Results from `try` `RS.collect` from each node, simplified if possible.
}
\description{
Force collect results to unlock error
}

