% Generated by roxygen2: do not edit by hand
\name{is.rsc}
\alias{is.rsc}
\title{Check if list stores Rserve connections}
\usage{
is.rsc(x, silent = TRUE)
}
\arguments{
\item{x}{a list}

\item{silent}{logical default TRUE, if FALSE then error are raised if list is not Rserve connections list.}
}
\value{
TRUE, FALSE or raise error if `silent=FALSE`.
}
\description{
Check if list stores Rserve connections
}

