% Generated by roxygen2: do not edit by hand
\name{rbindlapply}
\alias{rbindlapply}
\title{Row bind lapply results}
\usage{
rbindlapply(X, FUN, ..., use.names = fill, fill = FALSE, idcol = NULL)
}
\arguments{
\item{X}{vector (atomic or list) passed to `lapply`.}

\item{FUN}{function passed to `lapply`.}

\item{use.names}{logical passed to `rbindlist`.}

\item{fill}{logical passed to `rbindlist`.}

\item{idcol}{logical or  character passed to `rbindlist`.}

\item{\dots}{optional arguments to `FUN`.}
}
\description{
Wrapper on `rbindlist(lapply(...))`.
}

