% Generated by roxygen2: do not edit by hand
\name{rsc}
\alias{rsc}
\title{Connect to Rserve instances}
\usage{
rsc(port = 6311L, host = NULL, tls = FALSE, proxy.target = NULL,
  proxy.wait = TRUE)
}
\arguments{
\item{port}{integer vector of port numbers}

\item{host}{character scalar, soon will support vector.}

\item{tls}{logical, see `?RSclient::RS.connect` for details.}

\item{proxy.target}{logical, see `?RSclient::RS.connect` for details.}

\item{proxy.wait}{logical, see `?RSclient::RS.connect` for details.}
}
\value{
List of active connections to Rserve nodes.
}
\description{
A wrapper to `RS.connect` function using `lapply` over recycled input arguments. It also requires data.table package.
}

