% Generated by roxygen2: do not edit by hand
\name{rscl.assign}
\alias{rscl.assign}
\title{`RS.assign` for list of Rserve connections}
\usage{
rscl.assign(rscl = getOption("bigdatatable.rscl"), name, value, wait = TRUE,
  simplify = TRUE)
}
\arguments{
\item{rscl,}{lists of Rserve connections.}

\item{name}{character variable of name to be used in each node for new object.}

\item{value}{object to be assigned to each node.}

\item{wait}{logical default TRUE passed to `RS.eval`.}

\item{simplify}{logical, default TRUE, passed to underlying `sapply`.}
}
\value{
Results from `RS.assign` in `sapply`.
}
\description{
Wrapper on `RS.assign` for list of Rserve connections.
}

