% Generated by roxygen2: do not edit by hand
\name{rscl.clean}
\alias{rscl.clean}
\title{Force collect results to unlock error}
\usage{
rscl.clean(rscl = getOption("bigdatatable.rscl"))
}
\arguments{
\item{rscl}{list of connections to R nodes}
}
\value{
Results from `try` `RS.collect` from each node, simplified if possible.
}
\description{
Force collect results to unlock error
}

