% Generated by roxygen2: do not edit by hand
\name{rscl.collect}
\alias{rscl.collect}
\title{Collect results from Rserve connection list}
\usage{
rscl.collect(rscl = getOption("bigdatatable.rscl"), timeout = Inf,
  detail = FALSE, try = TRUE, simplify = TRUE)
}
\arguments{
\item{rscl}{list of connections to R nodes}

\item{timeout}{numeric passed to `RS.collect`.}

\item{detail}{logical passed to `RS.collect`.}

\item{try}{logical default TRUE, will wrap collection into `try` to finish collection from all nodes instead of aborting on error.}

\item{simplify}{logical, default TRUE, passed to underlying `sapply`.}
}
\value{
Results from `try` `RS.collect` from each node, simplified if possible.
}
\description{
Collect results from Rserve connection list
}

