% Generated by roxygen2: do not edit by hand
\name{rscl.connect}
\alias{rscl.connect}
\title{Connect to Rserve instances}
\usage{
rscl.connect(port = Sys.getenv("RSERVE_PORT", "6311"),
  host = Sys.getenv("RSERVE_HOST", "127.0.0.1"), tls = FALSE,
  proxy.target = NULL, proxy.wait = TRUE, pkgs = character())
}
\arguments{
\item{port}{integer vector of port numbers}

\item{host}{character scalar, soon will support vector.}

\item{tls}{logical, see `?RSclient::RS.connect` for details.}

\item{proxy.target}{logical, see `?RSclient::RS.connect` for details.}

\item{proxy.wait}{logical, see `?RSclient::RS.connect` for details.}

\item{pkgs}{character vector of packages to load after connect to node.}
}
\value{
List of active connections to Rserve nodes.
}
\description{
A wrapper to `RS.connect` function using `sapply` over recycled input arguments. It also requires data.table package.
}

