% Generated by roxygen2: do not edit by hand
\name{rscl.ls}
\alias{rscl.ls}
\title{`ls` on every node}
\usage{
rscl.ls(rscl, simplify = TRUE)
}
\arguments{
\item{rscl}{list of connections to R nodes}

\item{simplify}{logical default TRUE passed to `sapply`.}
}
\value{
`ls()` of `.GlobalEnv` from each node.
}
\description{
`ls` on every node
}

