% Generated by roxygen2: do not edit by hand
\name{rscl.require}
\alias{rscl.require}
\title{Require packages in list of Rserve instances}
\usage{
rscl.require(rscl, package, quietly = TRUE)
}
\arguments{
\item{rscl,}{lists of Rserve connections.}

\item{package}{character vector of packages to require on each of R node.}

\item{quietly}{logical defaul TRUE, no warning.}
}
\value{
Logical matrix.
}
\description{
A wrapper to `RS.eval`.
}

