% Generated by roxygen2: do not edit by hand
\name{[.big.data.table}
\alias{[.big.data.table}
\title{Subset from big.data.table}
\usage{
\method{[}{big.data.table}(x, ..., parallel = TRUE)
}
\arguments{
\item{x}{big.data.table object.}

\item{...}{arguments passed to each node `[.data.table` call: *i, j, by, keyby...*.}

\item{parallel}{logical if parallel *TRUE* (default) it will send expression to nodes using `wait=FALSE` and collect results afterward executing each node in parallel.}
}
\value{
data.table object.
}
\description{
Subset from big.data.table
}
\note{
Results from nodes are rbinded and the same call is evalated on combined results. That means the column names cannot be renamed or simplified to vector in `...` call. Use `[[.big.data.table` for deeper flexibility.
}

