% Generated by roxygen2: do not edit by hand
\name{[[.big.data.table}
\alias{[[.big.data.table}
\title{Extract from big.data.table}
\usage{
\method{[[}{big.data.table}(x, j, expr, lazy = TRUE, send = FALSE, i, ...,
  simplify = TRUE, rbind = TRUE, parallel = TRUE,
  outer.aggregate = getOption("bigdatatable.outer.aggregate", FALSE),
  .log = getOption("bigdatatable.log", FALSE))
}
\arguments{
\item{x}{big.data.table object.}

\item{j}{numeric scalar, if provided then all other arguments are ignored and subset behaves the same way as for data.table, but returns 0 length column.}

\item{expr}{expression to be evaluated on node where data.table objects are stored as `x` variable in `.GlobalEnv`.}

\item{lazy}{logical if TRUE then *expr* is substituted.}

\item{send}{logical, if TRUE submit expression appended with `TRUE` to not fetch potentially big results from provided *expr*, useful for data.table __set*__ or `:=` functions.}

\item{i}{numeric restrict expression to particular nodes}

\item{simplify}{logical passed to `bdt.eval`, affects the type of returned object.}

\item{rbind}{logical passed to `bdt.eval`, affects the type of returned object.}

\item{parallel}{logical if parallel *TRUE* (default) it will send expression to nodes using `wait=FALSE` and collect results afterward executing each node in parallel.}

\item{outer.aggregate}{logical or a function, if *TRUE* will able the same query to rbind of results from each node, should not be used with `.SD`, `.N`, etc. Also conflicts with filtering in `i`. Can be also a function taking first argument `x` rbinded data.table.}

\item{.log}{logical if *TRUE* then logging will be done using logR to postgres db.}

\item{\dots}{ignored.}
}
\value{
When using *j* arg the 0 length variable from underlying data is returned. Otherwise the results from expression evaluated as *lapply*. When using *rbind* or *simplify* the returned list be can simplified.
}
\description{
Extract from big.data.table
}

