% Generated by roxygen2: do not edit by hand
\name{bdt.eval}
\alias{bdt.eval}
\alias{bdt.eval.log}
\title{big.data.table evaluate on nodes}
\usage{
bdt.eval(x, expr, lazy = TRUE, send = FALSE, simplify = TRUE,
  rbind = TRUE, parallel = TRUE, outer.aggregate = FALSE, silent = TRUE,
  .log = getOption("bigdatatable.log", FALSE))

bdt.eval.log(x, expr, lazy = TRUE, send = FALSE, simplify = TRUE,
  rbind = TRUE, parallel = TRUE, outer.aggregate = FALSE, silent = TRUE,
  .log = getOption("bigdatatable.log", FALSE))
}
\arguments{
\item{x}{big.data.table.}

\item{expr}{expression.}

\item{lazy}{logical if TRUE then *expr* is substituted.}

\item{send}{logical, if TRUE submit expression appended with `TRUE` to not fetch potentially big results from provided *expr*, useful for data.table *set** or `:=` functions.}

\item{simplify}{logical if *TRUE* (default) it will simplify list of 1 length same type objects to vector.}

\item{rbind}{logical if *TRUE* (default) results are data.table they will be rbinded.}

\item{parallel}{logical if parallel *TRUE* (default) it will send expression to nodes using `wait=FALSE` and collect results afterward executing each node in parallel.}

\item{silent}{should be silently catched by `try` or `logR` if enabled.}

\item{.log}{logical if *TRUE* then logging will be done using logR to postgres db.}
}
\value{
Depending on *simplify, rbind* the results of evaluated expression.
}
\description{
Main engine for passing queries to nodes, control parallelization, rbinding or simplifing returned object. Allows to measure timing and verbose messages.
}
\details{
The `bdt.eval.log` function is wrapper on `bdt.eval` with `logR()` call, so use only when logR connected.
}

