% Generated by roxygen2: do not edit by hand
\docType{package}
\name{big.data.table-package}
\alias{bdt}
\alias{big.data.table}
\alias{big.data.table-package}
\alias{big.dt}
\title{big.data.table-package}
\description{
Leverage data.table onto set of Rserve nodes.
}
\details{
Use `as.big.data.table` methods to extract/load/assign/populate data to a list of R node connections.
Methods for `function` and `call` will be evaluated remotely, method for `data.table` will be evaluated locally and assigned to nodes.
}
\author{
Jan Gorecki
}

