% Generated by roxygen2: do not edit by hand
\name{core.data.table}
\alias{core.data.table}
\title{Core of data.table from nodes}
\usage{
core.data.table(x, var = "x")
}
\arguments{
\item{x}{big.data.table}

\item{var}{character scalar, variable name on remote node, usually 'x'.}
}
\value{
0 rows data.table class object, rbind of each 0L subsets.
}
\description{
Core of data.table from nodes
}

