% Generated by roxygen2: do not edit by hand
\name{is.big.data.table}
\alias{is.big.data.table}
\title{Test if object is big.data.table}
\usage{
is.big.data.table(x, check.nodes = FALSE)
}
\arguments{
\item{x}{R object.}

\item{check.nodes}{logical default FALSE, when TRUE it will validate that nodes have data.table as defined variable.}
}
\value{
For `check.nodes=FALSE` (default) a scalar logical if *x* inherits from *big.data.table*. For `check.nodes=TRUE` vector of results from expression `exists("x") && is.data.table(x)` on each node.
}
\description{
Test if object is big.data.table
}

