% Generated by roxygen2: do not edit by hand
\name{[[.big.data.table}
\alias{[[.big.data.table}
\title{Extract from big.data.table}
\usage{
\method{[[}{big.data.table}(x, j, expr, lazy = FALSE, send = FALSE, i, ...,
  simplify = TRUE, rbind = TRUE, parallel = TRUE)
}
\arguments{
\item{x}{big.data.table object.}

\item{j}{numeric scalar, if provided then all other arguments are ignored and subset behaves the same way as for data.table, but returns 0 length column.}

\item{expr}{expression to be evaluated on node where data.table objects are stored as `x` variable in `.GlobalEnv`.}

\item{lazy}{logical if TRUE then *expr* is substituted.}

\item{send}{logical, if TRUE submit expression appended with `TRUE` to not fetch potentially big results from provided *expr*, useful for data.table *set** or `:=` functions.}

\item{i}{numeric restrict expression to particular nodes}

\item{...}{ignored.}

\item{simplify}{logical passed to `bdt.eval`, affects the type of returned object.}

\item{rbind}{logical passed to `bdt.eval`, affects the type of returned object.}

\item{parallel}{logical if parallel *TRUE* (default) it will send expression to nodes using `wait=FALSE` and collect results afterward executing each node in parallel.}
}
\value{
When using *j* arg the 0 length variable from underlying data is returned. Otherwise the results from expression evaluated as `lapply*. When using *rbind* or *simplify* the returned list be can simplified.
}
\description{
Extract from big.data.table
}

